implement updateinfo => Errata import for spacewalk-repo-sync
updateinfo is converted into errata using the following assignments:

id: used in "advisory" and "advisory_name"
version: used in "advisory_rel".
         If 'from' contains 'suse' the version is also appended
         to "advisory" and "advisory_name", because suse re-use
         the id if newer version is available and only increase
         the version number.
         If version contains a number with dots, the number is
         converted to an integer.
type: is translated into spacewalk errata types and put into
      "advisory_type"
title: set to "synopsis"
release: set to "product"
issued: converted to db date and put into "issue_date"
references: bugzilla and CVE numbers are converted to the errata tables
description: set to "description"
The package list is evaluated and the packages are linked to the errata
reboot_suggested: this flag is added as keyword to errata
restart_suggested: this flag is added as keyword to errata
from: the errata table was enhanced. The column errata_from was added to
      have a place where the author/vendor could be stored.

The errata fields "topic" and "solution" are set to " "(empty) because
there is no equivalent in updateinfo.

+++ backend/satellite_tools/reposync.py
@@ -1,5 +1,6 @@
 #
 # Copyright (c) 2008--2011 Red Hat, Inc.
+# Copyright (c) 2010--2011 SUSE Linux Products GmbH
 #
 # This software is licensed to you under the GNU General Public License,
 # version 2 (GPLv2). There is NO WARRANTY for this software, express or
@@ -7,27 +8,32 @@
 # FOR A PARTICULAR PURPOSE. You should have received a copy of GPLv2
 # along with this software; if not, see
 # http://www.gnu.org/licenses/old-licenses/gpl-2.0.txt.
-# 
+#
 # Red Hat trademarks are not licensed under GPLv2. No permission is
 # granted to use or replicate Red Hat trademarks that are incorporated
-# in this software or its documentation. 
+# in this software or its documentation.
 #
 import sys, os, time
 import hashlib
+from datetime import datetime
+import traceback
 from optparse import OptionParser
 from spacewalk.server import rhnPackage, rhnSQL, rhnChannel, rhnPackageUpload
-from spacewalk.common import CFG, initCFG, rhnLog, rhnFault
+from spacewalk.common import CFG, initCFG, rhnLog, rhnFault, log_debug
 from spacewalk.common.checksum import getFileChecksum
-from spacewalk.server.importlib.importLib import IncompletePackage
+from spacewalk.common.rhn_mpm import InvalidPackageError
+from spacewalk.server.importlib.importLib import IncompletePackage, Erratum, Checksum, Bug, Keyword
 from spacewalk.server.importlib.packageImport import ChannelPackageSubscription
 from spacewalk.server.importlib.backendOracle import SQLBackend
+from spacewalk.common.rhn_mpm import InvalidPackageError
+from spacewalk.server.importlib.errataImport import ErrataImport
 from spacewalk.server import taskomatic
 
 default_log_location = '/var/log/rhn/reposync/'
 default_hash = 'sha256'
 
 class RepoSync:
-   
+
     parser = None
     type = None
     urls = None
@@ -48,7 +54,7 @@ class RepoSync:
             date = time.localtime()
             datestr = '%d.%02d.%02d-%02d:%02d:%02d' % (date.tm_year, date.tm_mon, date.tm_mday, date.tm_hour, date.tm_min, date.tm_sec)
             log_filename = options.channel_label + '-' +  datestr + '.log'
-           
+
         rhnLog.initLOG(default_log_location + log_filename)
         #os.fchown isn't in 2.4 :/
         os.system("chgrp apache " + default_log_location + log_filename)
@@ -101,6 +107,7 @@ class RepoSync:
         for url in self.urls:
             plugin = self.load_plugin()(url, self.channel_label)
             self.import_packages(plugin, url)
+            self.import_updates(plugin, url)
         if self.regen:
             taskomatic.add_to_repodata_queue_for_channel_package_subscription(
                 [self.channel_label], [], "server.app.yumreposync")
@@ -130,7 +137,174 @@ class RepoSync:
         mod = __import__('spacewalk.satellite_tools.repo_plugins', globals(), locals(), [name])
         submod = getattr(mod, name)
         return getattr(submod, "ContentSource")
-        
+
+    def import_updates(self, plug, url):
+      notices = plug.get_updates()
+      self.print_msg("Repo " + url + " has " + str(len(notices)) + " errata.")
+      if len(notices) > 0:
+        self.upload_updates(notices)
+
+    def upload_updates(self, notices):
+        batch = []
+        skipped_updates = 0
+        typemap = {
+                  'security'    : 'Security Advisory',
+                  'recommended' : 'Bug Fix Advisory',
+                  'bugfix'      : 'Bug Fix Advisory',
+                  'optional'    : 'Product Enhancement Advisory',
+                  'feature'     : 'Product Enhancement Advisory',
+                  'enhancement' : 'Product Enhancement Advisory'
+                  }
+        for notice in notices:
+            notice = self.fix_notice(notice)
+            existing_errata = self.get_errata(notice['update_id'])
+
+            e = Erratum()
+            e['errata_from']   = notice['from']
+            e['advisory']      = notice['update_id']
+            e['advisory_name'] = notice['update_id']
+            e['advisory_rel']  = notice['version']
+            e['advisory_type'] = typemap.get(notice['type'], 'Product Enhancement Advisory')
+            e['product']       = notice['release']
+            e['description']   = notice['description']
+            e['synopsis']      = notice['title']
+            e['topic']         = ' '
+            e['solution']      = ' '
+            e['issue_date']    = self._to_db_date(notice['issued'])
+            if notice['updated']:
+                e['update_date']   = self._to_db_date(notice['updated'])
+            else:
+                e['update_date']   = self._to_db_date(notice['issued'])
+            e['org_id']        = self.channel['org_id']
+            e['notes']         = ''
+            e['refers_to']     = ''
+            e['channels']      = []
+            e['packages']      = []
+            e['files']         = []
+            if existing_errata:
+                e['channels'] = existing_errata['channels']
+                e['packages'] = existing_errata['packages']
+            e['channels'].append({'label':self.channel_label})
+
+            for pkg in notice['pkglist'][0]['packages']:
+                param_dict = {
+                             'name'          : pkg['name'],
+                             'version'       : pkg['version'],
+                             'release'       : pkg['release'],
+                             'arch'          : pkg['arch'],
+                             'channel_label' : self.channel_label
+                             }
+                if pkg['epoch'] is None or pkg['epoch'] == '':
+                    epochStatement = "is NULL"
+                else:
+                    epochStatement = "= :epoch"
+                    param_dict['epoch'] = pkg['epoch']
+                if self.channel['org_id']:
+                    param_dict['org_id'] = self.channel['org_id']
+                    orgStatement = "= :org_id"
+                else:
+                    orgStatement = "is NULL"
+
+                h = rhnSQL.prepare("""
+                    select p.id, c.checksum, c.checksum_type
+                      from rhnPackage p
+                      join rhnPackagename pn on p.name_id = pn.id
+                      join rhnpackageevr pevr on p.evr_id = pevr.id
+                      join rhnpackagearch pa on p.package_arch_id = pa.id
+                      join rhnArchType at on pa.arch_type_id = at.id
+                      join rhnChecksumView c on p.checksum_id = c.id
+                      join rhnChannelPackage cp on p.id = cp.package_id
+                      join rhnChannel ch on cp.channel_id = ch.id
+                     where pn.name = :name
+                       and p.org_id %s
+                       and pevr.version = :version
+                       and pevr.release = :release
+                       and pa.label = :arch
+                       and pevr.epoch %s
+                       and at.label = 'rpm'
+                       and ch.label = :channel_label
+                """ % (orgStatement, epochStatement))
+                apply(h.execute, (), param_dict)
+                cs = h.fetchone_dict() or None
+
+                if not cs:
+                    if param_dict.has_key('epoch'):
+                        log_debug(1, "No cheksum found for %s-%s:%s-%s.%s. Skipping Package" % (param_dict['name'],
+                                                                                                param_dict['epoch'],
+                                                                                                param_dict['version'],
+                                                                                                param_dict['release'],
+                                                                                                param_dict['arch']
+                                                                                               ))
+                    else:
+                        log_debug(1, "No cheksum found for %s-%s-%s.%s. Skipping Package" % (param_dict['name'],
+                                                                                             param_dict['version'],
+                                                                                             param_dict['release'],
+                                                                                             param_dict['arch'],
+                                                                                            ))
+                    continue
+
+                newpkgs = []
+                for oldpkg in e['packages']:
+                    if oldpkg['package_id'] != cs['id']:
+                        newpkgs.append(oldpkg)
+
+                package = IncompletePackage()
+                for k in pkg.keys():
+                    package[k] = pkg[k]
+                package['epoch'] = pkg.get('epoch', '')
+                package['org_id'] = self.channel['org_id']
+
+                package['checksums'] = {cs['checksum_type'] : cs['checksum']}
+                package['checksum_type'] = cs['checksum_type']
+                package['checksum'] = cs['checksum']
+
+                package['package_id'] = cs['id']
+                newpkgs.append(package)
+
+                e['packages'] = newpkgs
+
+            if len(e['packages']) == 0:
+                skipped_updates = skipped_updates + 1
+                continue
+
+            e['keywords'] = []
+            if notice['reboot_suggested']:
+                kw = Keyword()
+                kw.populate({'keyword':'reboot_suggested'})
+                e['keywords'].append(kw)
+            if notice['restart_suggested']:
+                kw = Keyword()
+                kw.populate({'keyword':'restart_suggested'})
+                e['keywords'].append(kw)
+            e['bugs'] = []
+            e['cve'] = []
+            if notice['references']:
+                bzs = filter(lambda r: r['type'] == 'bugzilla', notice['references'])
+                if len(bzs):
+                    tmp = {}
+                    for bz in bzs:
+                        if bz['id'] not in tmp:
+                            bug = Bug()
+                            bug.populate({'bug_id' : bz['id'], 'summary' : bz['title'], 'href' : bz['href']})
+                            e['bugs'].append(bug)
+                            tmp[bz['id']] = None
+                cves = filter(lambda r: r['type'] == 'cve', notice['references'])
+                if len(cves):
+                    tmp = {}
+                    for cve in cves:
+                        if cve['id'] not in tmp:
+                            e['cve'].append(cve['id'])
+                            tmp[cve['id']] = None
+            e['locally_modified'] = None
+            batch.append(e)
+
+        if skipped_updates > 0:
+            self.print_msg("%d errata skipped because of empty package list." % skipped_updates)
+        backend = SQLBackend()
+        importer = ErrataImport(batch, backend)
+        importer.run()
+        self.regen = True
+
     def import_packages(self, plug, url):
         packages = plug.list_packages()
         to_process = []
@@ -168,7 +342,7 @@ class RepoSync:
         is_non_local_repo = (url.find("file://") < 0)
         # try/except/finally doesn't work in python 2.4 (RHEL5), so here's a hack
         def finally_remove(path):
-            if is_non_local_repo and path:
+            if is_non_local_repo and path and os.path.exists(path):
                 os.remove(path)
 
         for (index, what) in enumerate(to_process):
@@ -199,7 +373,7 @@ class RepoSync:
             getFileChecksum(checksum_type, filename=abspath) == checksum):
             return 1
         return 0
-    
+
     def upload_package(self, package):
         rel_package_path = rhnPackageUpload.relative_path_from_header(
                 package.header, self.channel['org_id'],
@@ -223,7 +397,7 @@ class RepoSync:
         package['arch'] = pack.arch
         package['checksum'] = pack.checksum
         package['checksum_type'] = pack.checksum_type
-        package['channels']  = [{'label':self.channel_label, 
+        package['channels']  = [{'label':self.channel_label,
                                  'id':self.channel['id']}]
         package['org_id'] = self.channel['org_id']
         try:
@@ -261,6 +435,80 @@ class RepoSync:
 
     def short_hash(self, str):
         return hashlib.new(default_hash, str).hexdigest()[0:8]
+
+    def _to_db_date(self, date):
+        ret = ""
+        if date.isdigit():
+          ret = datetime.fromtimestamp(float(date)).isoformat(' ')
+        else:
+          # we expect to get ISO formated date
+          ret = date
+        return ret
+
+    def fix_notice(self, notice):
+        if "." in notice['version']:
+            new_version = 0
+            for n in notice['version'].split('.'):
+                new_version = (new_version + int(n)) * 100
+            notice['version'] = new_version / 100
+        if "suse" in notice['from'].lower():
+            # suse style; we need to append the version to id
+            notice['update_id'] = notice['update_id'] + '-' + notice['version']
+        return notice
+
+    def get_errata(self, update_id):
+        h = rhnSQL.prepare("""select
+            e.id, e.advisory, e.advisory_name, e.advisory_rel
+            from rhnerrata e
+            where e.advisory = :name
+        """)
+        h.execute(name=update_id)
+        ret = h.fetchone_dict() or None
+        if not ret:
+            return None
+
+        h = rhnSQL.prepare("""select distinct c.label
+            from rhnchannelerrata ce
+            join rhnchannel c on c.id = ce.channel_id
+            where ce.errata_id = :eid
+        """)
+        h.execute(eid=ret['id'])
+        channels = h.fetchall_dict() or []
+
+        ret['channels'] = channels
+        ret['packages'] = []
+
+        h = rhnSQL.prepare("""
+            select p.id as package_id,
+                   pn.name,
+                   pevr.epoch,
+                   pevr.version,
+                   pevr.release,
+                   pa.label as arch,
+                   p.org_id,
+                   cv.checksum,
+                   cv.checksum_type
+              from rhnerratapackage ep
+              join rhnpackage p on p.id = ep.package_id
+              join rhnpackagename pn on pn.id = p.name_id
+              join rhnpackageevr pevr on pevr.id = p.evr_id
+              join rhnpackagearch pa on pa.id = p.package_arch_id
+              join rhnchecksumview cv on cv.id = p.checksum_id
+             where ep.errata_id = :eid
+        """)
+        h.execute(eid=ret['id'])
+        packages = h.fetchall_dict() or []
+        for pkg in packages:
+            ipackage = IncompletePackage()
+            for k in pkg.keys():
+                ipackage[k] = pkg[k]
+            ipackage['epoch'] = pkg.get('epoch', '')
+
+            ipackage['checksums'] = {ipackage['checksum_type'] : ipackage['checksum']}
+            ret['packages'].append(ipackage)
+
+        return ret
+
 
 class ContentPackage:
